factor out fwd stuff to drop headers dependencies
